moving validate schema logic into openwrt stuff
