+++ sync/managers.py
@@ -2,7 +2,7 @@
 # pylint: disable=bare-except
 import traceback
 import os
-from collections import OrderedDict
+
 from . import registrar
 
 
@@ -31,7 +31,6 @@ def validate_settings(settings):
     """
     Validate the settings
     """
-    validate_schema(settings)
     for manager in registrar.managers:
         manager.validate_settings(settings)
 
@@ -85,44 +84,3 @@ def create_settings_in_tmpdir(settings_filename, tmpdir, tmpdir_delete):
         file.close()
 
     return 0
-
-def validate_schema(settings):
-    """
-    Eventually this should validate the schema against mfw_schema, at which point the recursion can probably be removed
-    """
-    bad_att_locations = []
-
-    schema_recurse(settings, bad_att_locations)
-
-    if len(bad_att_locations) > 0:
-        raise Exception("Schema Validation: Bad attributes found. JSON Locations: %s " % bad_att_locations)
-
-
-def schema_recurse(currentItem, bad_attr_locations, itemParents=['root']):
-    """
-    This function currently recurses the entire json schema for attribute names of 'output', 'result', or 'error' and raises an exception if found
-
-    """
-    # print("Starting schema scan on: %s" % itemParents)
-    bad_attributes = ['output', 'results', 'error']
-
-    if isinstance(currentItem, OrderedDict):
-        iterator = currentItem.items()
-    elif isinstance(currentItem, list):
-        iterator = enumerate(currentItem) 
-    else:
-        # print("Attribute: %s is non iterable" % currentItem)
-        return
-
-    for k, v in iterator:
-        # print("Schema Validation: Verifying: %s type: %s" % (k, type(v)))
-
-        if k in bad_attributes:
-            dataLocation = str('/'.join(itemParents)) + '/' +  k
-            print("Bad JSON data found: %s " % dataLocation)
-            bad_attr_locations.append(dataLocation)
-
-        if isinstance(v, OrderedDict) or isinstance(v, list):
-            itemParents.append(str(k))
-            schema_recurse(v, bad_attr_locations, itemParents)
-            del itemParents[-1]
