+++ sync/openwrt/settings_manager.py
@@ -4,6 +4,7 @@ import os
 import json
 import shutil
 from sync import registrar
+from collections import OrderedDict
 
 # This class is responsible for writing /etc/config/network
 # based on the settings object passed from sync-settings
@@ -25,6 +26,7 @@ class SettingsManager:
 
     def validate_settings(self, settings):
         """validates settings"""
+        validate_schema(settings)
         pass
 
     def create_settings(self, settings, prefix, delete_list, filepath):
@@ -59,3 +61,45 @@ class SettingsManager:
         print("%s: Wrote %s" % (self.__class__.__name__, filename))
 
 registrar.register_manager(SettingsManager())
+
+
+def validate_schema(settings):
+    """
+    Eventually this should validate the schema against mfw_schema, at which point the recursion can probably be removed
+    """
+    bad_att_locations = []
+
+    schema_recurse(settings, bad_att_locations)
+
+    if len(bad_att_locations) > 0:
+        raise Exception("Schema Validation: Bad attributes found. JSON Locations: %s " % bad_att_locations)
+
+
+def schema_recurse(currentItem, bad_attr_locations, itemParents=['root']):
+    """
+    This function currently recurses the entire json schema for attribute names of 'output', 'result', or 'error' and raises an exception if found
+
+    """
+    print("Starting schema scan on: %s" % itemParents)
+    bad_attributes = ['output', 'results', 'error']
+
+    if isinstance(currentItem, OrderedDict):
+        iterator = currentItem.items()
+    elif isinstance(currentItem, list):
+        iterator = enumerate(currentItem) 
+    else:
+        # print("Attribute: %s is non iterable" % currentItem)
+        return
+
+    for k, v in iterator:
+        # print("Schema Validation: Verifying: %s type: %s" % (k, type(v)))
+
+        if k in bad_attributes:
+            dataLocation = str('/'.join(itemParents)) + '/' +  k
+            print("Bad JSON data found: %s " % dataLocation)
+            bad_attr_locations.append(dataLocation)
+
+        if isinstance(v, OrderedDict) or isinstance(v, list):
+            itemParents.append(str(k))
+            schema_recurse(v, bad_attr_locations, itemParents)
+            del itemParents[-1]
